Added orphan crop cleanup functionality to the save event. Now when an element with saved crops gets removed, the dir and files are also deleted.

+++ core/components/fredcropper/model/fredcropper/src/Endpoint/Ajax/SaveCrop.php
@@ -146,7 +146,9 @@ class SaveCrop extends Endpoint {
 
     protected function deleteOldCrop() {
         if (file_exists($this->elementDir.$this->oldFilename)) {
-            unlink($this->elementDir.$this->oldFilename);
+            if(is_file($this->elementDir.$this->oldFilename)) {
+                unlink($this->elementDir . $this->oldFilename);
+            }
         }
 
     }
